refactor: rename use-auth.ts to useAuth.ts for consistent naming
- Follow consistent naming convention for custom hooks
- Update all import statements to use new filename
- No functional changes, only filename and imports
